+++ encoding/features/__init__.py
@@ -1,6 +1,6 @@
 from .language_model import LanguageModelFeatureExtractor
 from .speech_model import SpeechFeatureExtractor
-from .simple_features import WordRateFeatureExtractor
+from .lowlevel_features import WordRateFeatureExtractor
 from .embeddings import StaticEmbeddingFeatureExtractor
 from .FIR_expander import FIR
 from .factory import FeatureExtractorFactory
